+++ modules/msp/apm/service/components/transaction-rpc-slow/metric_table_filter/provider.go
@@ -68,11 +68,11 @@ func (f *FrontendConditions) convertToTransactionFilter() slow_transaction.SlowT
 	}
 	d, err := time.ParseDuration(fmt.Sprintf("%d%s", f.Duration[0].Timer, f.Duration[0].Unit))
 	if err == nil {
-		slowTransactionFilter.MinDuration = int64(d)
+		slowTransactionFilter.MinDuration = float64(d)
 	}
 	d, err = time.ParseDuration(fmt.Sprintf("%d%s", f.Duration[1].Timer, f.Duration[1].Unit))
 	if err == nil {
-		slowTransactionFilter.MaxDuration = int64(d)
+		slowTransactionFilter.MaxDuration = float64(d)
 	}
 	return slowTransactionFilter
 }
